// libs/helpers/src/types.ts
export interface IStoryFragmentCompositorHooks {
  processRead: Function;
  belief?: Function;
  identifyAs?: Function;
  youtube?: Function;
  toggle?: Function;
  updateEventStream?: Function;
  resourcePayload?: any;
  templates?: any;
  setScrollToPane?: Function;
  Link?: Function;
}
// ...
export interface IStoryFragmentId {
  id: string;
  title: string;
  slug: string;
  tractStackId: string;
  tractStackTitle: string;
  tractStackSlug: string;
  home: string;
  isContextPane?: boolean;
  isBuilderPreview?: boolean;
  viewportKey?: string;
  paneId?: string;
  paneTitle?: string;
  paneSlug?: string;
}